requirements.txt
numpy
pandas
datasets
tqdm
openai
python-dotenv
sentence-transformers
tiktoken
FlagEmbedding